feat: first look on working version
