+++ src/createNativeStackNavigator.tsx
@@ -196,10 +196,6 @@ class StackView extends React.Component<Props> {
       hideShadow,
       largeTitle,
       largeTitleHideShadow,
-      screenOrientation,
-      statusBarAnimation,
-      statusBarHidden,
-      statusBarStyle,
       title,
       translucent,
     } = options;
@@ -231,10 +227,6 @@ class StackView extends React.Component<Props> {
       largeTitleFontSize: headerLargeTitleStyle?.fontSize,
       largeTitleFontWeight: headerLargeTitleStyle?.fontWeight,
       largeTitleHideShadow: largeTitleHideShadow || headerLargeTitleHideShadow,
-      screenOrientation,
-      statusBarAnimation,
-      statusBarHidden,
-      statusBarStyle,
       title,
       titleColor: headerTitleStyle?.color || headerTintColor,
       titleFontFamily: headerTitleStyle?.fontFamily,
@@ -384,6 +376,10 @@ class StackView extends React.Component<Props> {
             ? true
             : options.gestureEnabled
         }
+        screenOrientation={options.screenOrientation}
+        statusBarAnimation={options.statusBarAnimation}
+        statusBarHidden={options.statusBarHidden}
+        statusBarStyle={options.statusBarStyle}
         onAppear={() => this.onAppear(route, descriptor)}
         onWillAppear={() => options?.onWillAppear?.()}
         onWillDisappear={() => options?.onWillDisappear?.()}
