+++ src/types.tsx
@@ -96,6 +96,18 @@ export interface ScreenProps extends ViewProps {
    */
   replaceAnimation?: ScreenReplaceTypes;
   /**
+   * @description Controls in which orientation should the screen appear.
+   * @type "default" - on iOS, it resolves to "all" without "portrait_down" and on Android it lets the system decide the best orientation
+   * @type "all" – all orientations are permitted
+   * @type "portrait" – portrait orientations are permitted
+   * @type "portrait_up" – right-side portrait orientation is permitted
+   * @type "portrait_down" – upside-down portrait orientation is permitted
+   * @type "landscape" – landscape orientations are permitted
+   * @type "landscape_left" – landscape-left orientation is permitted
+   * @type "landscape_right" – landscape-right orientation is permitted
+   */
+  screenOrientation?: ScreenOrientationTypes;
+  /**
    * @description Allows for the customization of how the given screen should appear/dissapear when pushed or popped at the top of the stack. The following values are currently supported:
    *  @type "default" – uses a platform default animation
    *  @type "fade" – fades screen in or out
@@ -115,6 +127,21 @@ export interface ScreenProps extends ViewProps {
    * @type "formSheet" – will use "UIModalPresentationFormSheet" modal style on iOS and will fallback to "modal" on Android.
    */
   stackPresentation?: StackPresentationTypes;
+  /**
+   * @host (iOS only)
+   * @description Sets the status bar animation (similar to the `StatusBar` component). Requires enabling (or deleting) `View controller-based status bar appearance` in your Info.plist file. Defaults to `fade`.
+   */
+  statusBarAnimation?: 'none' | 'fade' | 'slide';
+  /**
+   * @host (iOS only)
+   * @description When set to true, the status bar for this screen is hidden. Requires enabling (or deleting) `View controller-based status bar appearance` in your Info.plist file. Defaults to `false`.
+   */
+  statusBarHidden?: boolean;
+  /**
+   * @host (iOS only)
+   * @description Sets the status bar color (similar to the `StatusBar` component). Requires enabling (or deleting) `View controller-based status bar appearance` in your Info.plist file. Defaults to `auto`.
+   */
+  statusBarStyle?: 'inverted' | 'auto' | 'light' | 'dark';
 }
 
 export interface ScreenContainerProps extends ViewProps {
@@ -219,33 +246,6 @@ export interface ScreenStackHeaderConfigProps extends ViewProps {
    * @description Boolean that allows for disabling drop shadow under navigation header when the edge of any scrollable content reaches the matching edge of the navigation bar.
    */
   largeTitleHideShadow?: boolean;
-  /**
-   * @description Controls in which orientation should the screen appear.
-   * @type "default" - on iOS, it resolves to "all" without "portrait_down" and on Android it lets the system decide the best orientation
-   * @type "all" – all orientations are permitted
-   * @type "portrait" – portrait orientations are permitted
-   * @type "portrait_up" – right-side portrait orientation is permitted
-   * @type "portrait_down" – upside-down portrait orientation is permitted
-   * @type "landscape" – landscape orientations are permitted
-   * @type "landscape_left" – landscape-left orientation is permitted
-   * @type "landscape_right" – landscape-right orientation is permitted
-   */
-  screenOrientation?: ScreenOrientationTypes;
-  /**
-   * @host (iOS only)
-   * @description Sets the status bar animation (similar to the `StatusBar` component). Requires enabling (or deleting) `View controller-based status bar appearance` in your Info.plist file. Defaults to `fade`.
-   */
-  statusBarAnimation?: 'none' | 'fade' | 'slide';
-  /**
-   * @host (iOS only)
-   * @description When set to true, the status bar for this screen is hidden. Requires enabling (or deleting) `View controller-based status bar appearance` in your Info.plist file. Defaults to `false`.
-   */
-  statusBarHidden?: boolean;
-  /**
-   * @host (iOS only)
-   * @description Sets the status bar color (similar to the `StatusBar` component). Requires enabling (or deleting) `View controller-based status bar appearance` in your Info.plist file. Defaults to `auto`.
-   */
-  statusBarStyle?: 'inverted' | 'auto' | 'light' | 'dark';
   /**
    * @description String that representing screen title that will get rendered in the middle section of the header. On iOS the title is centered on the header while on Android it is aligned to the left and placed next to back button (if one is present).
    */
